Refine arbitrage opportunity model
