+++ Models/ArbOpportunity.swift
@@ -1,1 +1,19 @@
+import Foundation
 
+public struct ArbOpportunity: Identifiable, Hashable, Sendable {
+    public let id = UUID()
+    public let tokenPair: String
+    public let buyExchange: String
+    public let sellExchange: String
+    public let buyPrice: Double
+    public let sellPrice: Double
+    public let timestamp: Date
+
+    /// Absolute profit between the buy and sell legs.
+    public var profit: Double { sellPrice - buyPrice }
+
+    /// Percentage spread expressed as a fraction (0.01 = 1%).
+    public var spread: Double { (sellPrice - buyPrice) / buyPrice }
+}
+
+public typealias ArbitrageOpportunity = ArbOpportunity
